refactor: comment - controller, repository, service test

+++ server/src/main/java/com/matzip/server/domain/review/service/ReviewService.java
@@ -2,6 +2,7 @@ package com.matzip.server.domain.review.service;
 
 import com.matzip.server.domain.image.service.ImageService;
 import com.matzip.server.domain.review.dto.ReviewDto;
+import com.matzip.server.domain.review.exception.AccessBlockedOrDeletedReviewException;
 import com.matzip.server.domain.review.exception.ReviewChangeByAnonymousException;
 import com.matzip.server.domain.review.exception.ReviewNotFoundException;
 import com.matzip.server.domain.review.model.Review;
@@ -29,15 +30,16 @@ public class ReviewService {
     @Transactional
     public ReviewDto.Response postReview(Long myId, ReviewDto.PostRequest postRequest) {
         User me = userRepository.findMeById(myId);
-        Review review = new Review(me, postRequest);
+        Review review = reviewRepository.save(new Review(me, postRequest));
         imageService.uploadReviewImages(me, review, postRequest.getImages());
         return ReviewDto.Response.of(review, me);
     }
 
     public ReviewDto.Response fetchReview(Long myId, Long reviewId) {
         User me = userRepository.findMeById(myId);
-        return ReviewDto.Response.of(
-                reviewRepository.findById(reviewId).orElseThrow(() -> new ReviewNotFoundException(reviewId)), me);
+        Review review = reviewRepository.findById(reviewId).orElseThrow(() -> new ReviewNotFoundException(reviewId));
+        if (review.isBlocked() || review.isDeleted()) throw new AccessBlockedOrDeletedReviewException(reviewId);
+        return ReviewDto.Response.of(review, me);
     }
 
     public Slice<ReviewDto.Response> searchReviews(Long myId, ReviewDto.SearchRequest searchRequest) {
@@ -50,14 +52,15 @@ public class ReviewService {
     public ReviewDto.Response patchReview(Long myId, Long reviewId, ReviewDto.PatchRequest patchRequest) {
         User me = userRepository.findMeById(myId);
         Review review = reviewRepository.findById(reviewId).orElseThrow(() -> new ReviewNotFoundException(reviewId));
+        if (review.isBlocked() || review.isDeleted()) throw new AccessBlockedOrDeletedReviewException(reviewId);
         if (review.getUser() != me)
             throw new ReviewChangeByAnonymousException();
         if (Optional.ofNullable(patchRequest.getContent()).isPresent())
             review.setContent(patchRequest.getContent());
-        if (Optional.ofNullable(patchRequest.getNewImages()).isPresent())
-            imageService.uploadReviewImages(me, review, patchRequest.getNewImages());
         if (Optional.ofNullable(patchRequest.getOldUrls()).isPresent())
             imageService.deleteReviewImages(review, patchRequest.getOldUrls());
+        if (Optional.ofNullable(patchRequest.getNewImages()).isPresent())
+            imageService.uploadReviewImages(me, review, patchRequest.getNewImages());
         if (Optional.ofNullable(patchRequest.getRating()).isPresent())
             review.setRating(patchRequest.getRating());
         return ReviewDto.Response.of(review, me);
@@ -67,6 +70,7 @@ public class ReviewService {
     public void deleteReview(Long myId, Long reviewId) {
         User me = userRepository.findMeById(myId);
         Review review = reviewRepository.findById(reviewId).orElseThrow(() -> new ReviewNotFoundException(reviewId));
+        if (review.isBlocked() || review.isDeleted()) throw new AccessBlockedOrDeletedReviewException(reviewId);
         if (review.getUser() != me)
             throw new ReviewChangeByAnonymousException();
         review.delete();
